# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Graph Search API
  version: 1.0.0
  description: >
    A FastAPI-powered code knowledge graph browser. Index source code
    repositories into a navigable graph of symbols and relationships,
    explore the graph visually, run searches, and chat with an
    AI assistant that understands your codebase.
servers:
  - url: /
...
paths:
...
  /api/projects/open:
    post:
      operationId: openProject
      tags: [Projects]
      summary: Open or switch to a project directory
      description: >
        If `apollo.json` exists with `initial_index_completed=true`, returns
        a project info payload with `needs_bootstrap=false`. Otherwise
        signals the bootstrap wizard. Refuses to open a directory that
        sits inside another already-initialized project.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path:
                  type: string
                  description: Absolute path to the project directory.
      responses:
        "200":
          description: Project info
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/projects/init:
    post:
      operationId: initProject
      tags: [Projects]
      summary: Initialize a new project with custom filters
      description: Triggered when the bootstrap wizard submits its filter selections.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path:
                  type: string
                filters:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: Project info
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/projects/filters:
    put:
      operationId: updateProjectFilters
      tags: [Projects]
      summary: Update inclusion/exclusion filters for the current project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [filters]
              properties:
                filters:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: Updated project info
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/projects/reprocess:
    post:
      operationId: reprocessProject
      tags: [Projects]
      summary: Trigger a full or incremental re-index of the open project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mode:
                  type: string
                  enum: [incremental, full]
                  default: incremental
      responses:
        "200":
          description: Job queued
          content:
            application/json:
              schema:
                type: object
                required: [status, mode, project_id]
                properties:
                  status:
                    type: string
                    enum: [queued]
                  mode:
                    type: string
                    enum: [incremental, full]
                  project_id:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/projects/leave:
    post:
      operationId: leaveProject
      tags: [Projects]
      summary: Remove the current project (deletes `_apollo/` and `_apollo_web/`)
      description: Requires `confirm=true`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [confirm]
              properties:
                confirm:
                  type: boolean
      responses:
        "200":
          description: Project removed
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    enum: [removed]
                  deleted:
                    type: array
                    items:
                      type: string
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/projects/tree:
    get:
      operationId: getProjectTree
      tags: [Projects]
      summary: Folder tree of the current project (for bootstrap wizard)
      parameters:
        - name: depth
          in: query
          schema:
            type: integer
            default: 3
            description: Maximum recursion depth.
      responses:
        "200":
          description: Tree node with `name`, `path`, `type`, `children`, and per-directory counts.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          $ref: "#/components/responses/BadRequest"

  # ── Realtime (WebSocket) ────────────────────────────────────────────
...
  /api/annotations:
    get:
      operationId: listAnnotations
      tags: [Annotations]
      summary: List all annotations (highlights, notes, bookmarks, tags)
      description: >
        Returns every annotation in the active project, newest first.
        Use this to give an AI assistant the user's own **notes** and
        **bookmarks** alongside graph data — each annotation's
        `target` field links it to either a file path or a graph node ID,
        which can then be expanded with `/api/node/{node_id}` or
        `/api/neighbors/{node_id}` for relationship context.
      parameters:
        - name: type
          in: query
          required: false
          description: Optional type filter.
          schema:
            type: string
            enum: [highlight, bookmark, note, tag]
      responses:
        "200":
          description: List of annotations
          content:
            application/json:
              schema:
                type: object
                required: [annotations]
                properties:
                  annotations:
                    type: array
                    items:
                      $ref: "#/components/schemas/Annotation"
  /api/annotations/create:
    post:
      operationId: createAnnotation
      tags: [Annotations]
      summary: Create a new annotation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnotationCreate"
      responses:
        "200":
          description: Created annotation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Annotation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/annotations/by-target:
    get:
      operationId: annotationsByTarget
      tags: [Annotations]
      summary: Find annotations attached to a file path or graph node
      description: >
        Provide exactly one of `file` or `node`. Useful for joining
        notes to graph context — e.g. fetch every note attached to
        a node returned by `search_graph`.
      parameters:
        - name: file
          in: query
          required: false
          schema:
            type: string
          description: Project-relative file path.
        - name: node
          in: query
          required: false
          schema:
            type: string
          description: Graph node ID, e.g. `func::src/main.py::main`.
      responses:
        "200":
          description: Matching annotations
          content:
            application/json:
              schema:
                type: object
                required: [annotations]
                properties:
                  annotations:
                    type: array
                    items:
                      $ref: "#/components/schemas/Annotation"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/annotations/by-tag:
    get:
      operationId: annotationsByTag
      tags: [Annotations]
      summary: Find annotations carrying a given tag
      parameters:
        - name: tag
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Matching annotations
          content:
            application/json:
              schema:
                type: object
                required: [annotations]
                properties:
                  annotations:
                    type: array
                    items:
                      $ref: "#/components/schemas/Annotation"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/annotations/collections:
    get:
      operationId: listAnnotationCollections
      tags: [Annotations]
      summary: List annotation collections
      responses:
        "200":
          description: All collections in the active project
          content:
            application/json:
              schema:
                type: object
                required: [collections]
                properties:
                  collections:
                    type: array
                    items:
                      $ref: "#/components/schemas/AnnotationCollection"
    post:
      operationId: createAnnotationCollection
      tags: [Annotations]
      summary: Create an annotation collection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
                annotation_ids:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Created collection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnotationCollection"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/annotations/collections/{collection_id}:
    delete:
      operationId: deleteAnnotationCollection
      tags: [Annotations]
      summary: Delete an annotation collection
      parameters:
        - name: collection_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deletion confirmation
          content:
            application/json:
              schema:
                type: object
                required: [deleted]
                properties:
                  deleted:
                    type: string
        "404":
          $ref: "#/components/responses/NotFound"
  /api/annotations/{annotation_id}:
    get:
      operationId: getAnnotation
      tags: [Annotations]
      summary: Get a single annotation
      parameters:
        - name: annotation_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Annotation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Annotation"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateAnnotation
      tags: [Annotations]
      summary: Update an annotation
      parameters:
        - name: annotation_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnotationUpdate"
      responses:
        "200":
          description: Updated annotation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Annotation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteAnnotation
      tags: [Annotations]
      summary: Delete an annotation
      parameters:
        - name: annotation_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deletion confirmation
          content:
            application/json:
              schema:
                type: object
                required: [deleted]
                properties:
                  deleted:
                    type: string
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Local AI tools (PLAN_MORE_LOCAL_AI_FUNCTIONS) ─────────────────────
  # Each endpoint below mirrors a tool in `chat/local_tools.py` so the
  # human-facing UI has parity with the AI agent's view. All read-only.
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [status_code, error, detail]
      properties:
        status_code:
          type: integer
        error:
          type: string
        detail:
          oneOf:
            - type: string
            - type: object
            - type: array

    # ── Filesystem ────────────────────────────────────────────────────
...
    HighlightRange:
      type: object
      required: [start_line, end_line]
      properties:
        start_line:
          type: integer
        end_line:
          type: integer
        start_col:
          type: integer
        end_col:
          type: integer
    Annotation:
      type: object
      required: [id, type, target, created_at]
      properties:
        id:
          type: string
          description: Unique annotation ID, e.g. `an::a1b2c3d4e5f60718`.
        type:
          type: string
          enum: [highlight, bookmark, note, tag]
        target:
          $ref: "#/components/schemas/AnnotationTarget"
        created_at:
          type: string
          format: date-time
        content:
          type:
            - string
            - "null"
          description: Free-form note text (markdown allowed).
        tags:
          type: array
          items:
            type: string
        color:
          type:
            - string
            - "null"
          enum: [red, yellow, green, blue, purple, gray, null]
        highlight_range:
          $ref: "#/components/schemas/HighlightRange"
        stale:
          type: boolean
          description: True if the target file or node no longer exists in the index.
        last_modified_at:
          type:
            - string
            - "null"
          format: date-time
    AnnotationCreate:
      type: object
      required: [type, target]
      properties:
        type:
          type: string
          enum: [highlight, bookmark, note, tag]
        target:
          $ref: "#/components/schemas/AnnotationTarget"
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        color:
          type: string
          enum: [red, yellow, green, blue, purple, gray]
        highlight_range:
          $ref: "#/components/schemas/HighlightRange"
    AnnotationUpdate:
      type: object
      description: Any subset of fields. Omitted fields are left unchanged.
      properties:
        type:
          type: string
          enum: [highlight, bookmark, note, tag]
        target:
          $ref: "#/components/schemas/AnnotationTarget"
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        color:
          type: string
          enum: [red, yellow, green, blue, purple, gray]
        highlight_range:
          $ref: "#/components/schemas/HighlightRange"
        stale:
          type: boolean
    AnnotationCollection:
      type: object
      required: [id, name, annotation_ids, created_at]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type:
            - string
            - "null"
        annotation_ids:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time

    # ── Local AI tools (PLAN_MORE_LOCAL_AI_FUNCTIONS) ─────────────────
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            status_code: 400
            error: Bad Request
            detail: The provided path is not a directory.
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            status_code: 404
            error: Not Found
            detail: The requested resource was not found.
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            status_code: 500
            error: Internal Server Error
            detail: An unexpected error occurred.